WIP: rename to userland, no more KUBECTL_PLUGINS_*
the environment variables in KUBECTL_PLUGINS_* basically aren't a thing
anymore, need to go away as described in
https://kubernetes.io/docs/tasks/extend-kubectl/kubectl-plugins/#flags-and-argument-handling

we need a different way to get $KUBECTL, $NAMESPACE, $TARGETUSER, and
the container name.  I've set some defaults that work for me, but with
the possible exception of `which kubectl` they are probably no good for
anyone else.